% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/age_group.R
\name{age_group}
\alias{age_group}
\title{Classify age group}
\usage{
age_group(x, droplevel = TRUE)
}
\arguments{
\item{x}{a vector of ages}
}
\value{
factor
}
\description{
classify the age group of the input age
}
\examples{
age_group(c(2,70,45,22,27))
}

