% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factor_level.R
\name{factor_level}
\alias{factor_level}
\title{user defined factor levels}
\usage{
factor_level(x, level = "alp", ...)
}
\arguments{
\item{x}{a vector which is a potential factor}

\item{level}{Valid levels for x, default is "alp".
User defined level character must match the number of levels of x, and must include all values of x.
If equals "alp",levels of x appears alphabetically;
if equals "asit",levels of x are ordered as they appeared in x.}

\item{...}{optional arguments to pass to the function factor()}
}
\value{
x as a factor with level orders specified
}
\description{
It turns input to a fator with factor levels order defined sepcifically
by user
}
\examples{
factor_level(iris$Species,level = c("versicolor","setosa","virginica"))
}

