% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/order_asit.R
\name{order_asit}
\alias{order_asit}
\title{Set levels of a factor to the order appeared in data}
\usage{
order_asit(x)
}
\arguments{
\item{x}{a vector}
}
\value{
x as a factor
}
\description{
Set levels of a factor to the order appeared in data
}
\examples{
order_asit(c("small","large","median","large","small"))
}

